+++ js/main.js
@@ -1,7 +1,7 @@
 import {buildAnnouncements} from './build-announcements.js';
-import {showCardOnPage} from './show-card.js';
+import {card} from './card.js';
 
 const NUMBER_OF_ANNOUNCEMENTS = 10;
 const listCards = buildAnnouncements(NUMBER_OF_ANNOUNCEMENTS);
 
-showCardOnPage(listCards[0]);
+card.showOnPage(listCards[0]);
